chore(core): Refactor options package names
