+++ pkg/core/options/options_test.go
@@ -36,463 +36,6 @@ var (
 	rootPath   = field.NewPath("root")
 )
 
-func TestValidateOptionSpec(t *testing.T) {
-	tests := []struct {
-		name    string
-		spec    *execution.OptionSpec
-		wantErr bool
-	}{
-		{
-			name: "nil config",
-			spec: nil,
-		},
-		{
-			name: "no options",
-			spec: &execution.OptionSpec{
-				Options: nil,
-			},
-		},
-		{
-			name: "duplicate option names",
-			spec: &execution.OptionSpec{
-				Options: []execution.Option{
-					{
-						Name:   "option",
-						Type:   execution.OptionTypeString,
-						String: &execution.StringOptionConfig{},
-					},
-					{
-						Name: "option",
-						Type: execution.OptionTypeSelect,
-						Select: &execution.SelectOptionConfig{
-							Values: []string{"a", "b"},
-						},
-					},
-				},
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			err := options.ValidateOptionSpec(tt.spec, rootPath).ToAggregate()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ValidateOptionSpec() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func TestValidateJobOption(t *testing.T) {
-	tests := []struct {
-		name    string
-		option  execution.Option
-		wantErr bool
-	}{
-		{
-			name: "invalid type",
-			option: execution.Option{
-				Type: "__invalid__",
-				Name: "opt",
-			},
-			wantErr: true,
-		},
-		{
-			name: "empty name",
-			option: execution.Option{
-				Type:   execution.OptionTypeString,
-				String: &execution.StringOptionConfig{},
-			},
-			wantErr: true,
-		},
-		{
-			name: "name has spaces",
-			option: execution.Option{
-				Type:   execution.OptionTypeString,
-				Name:   "option name",
-				String: &execution.StringOptionConfig{},
-			},
-			wantErr: true,
-		},
-		{
-			name: "name has unallowed characters",
-			option: execution.Option{
-				Type:   execution.OptionTypeString,
-				Name:   "<option_name>",
-				String: &execution.StringOptionConfig{},
-			},
-			wantErr: true,
-		},
-		{
-			name: "bool option",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Default: true,
-					Format:  execution.BoolOptionFormatTrueFalse,
-				},
-			},
-		},
-		{
-			name: "bool option missing fields",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{},
-			},
-			wantErr: true,
-		},
-		{
-			name: "bool option invalid format",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Format: "invalid",
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "bool option custom format empty values ok",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Format: execution.BoolOptionFormatCustom,
-				},
-			},
-		},
-		{
-			name: "bool option cannot be required",
-			option: execution.Option{
-				Type:     execution.OptionTypeBool,
-				Name:     "opt",
-				Required: true,
-				Bool: &execution.BoolOptionConfig{
-					Default: true,
-					Format:  execution.BoolOptionFormatTrueFalse,
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "bool option empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-			},
-			wantErr: true,
-		},
-		{
-			name: "bool option with non-bool config",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Format: execution.BoolOptionFormatTrueFalse,
-				},
-				String: &execution.StringOptionConfig{},
-			},
-			wantErr: true,
-		},
-		{
-			name: "string option",
-			option: execution.Option{
-				Type: execution.OptionTypeString,
-				Name: "opt",
-				String: &execution.StringOptionConfig{
-					Default:    "hello",
-					TrimSpaces: true,
-				},
-			},
-		},
-		{
-			name: "string option require default value",
-			option: execution.Option{
-				Type:     execution.OptionTypeString,
-				Name:     "opt",
-				Required: true,
-				String:   &execution.StringOptionConfig{},
-			},
-		},
-		{
-			name: "string option require default value after trimming",
-			option: execution.Option{
-				Type:     execution.OptionTypeString,
-				Name:     "opt",
-				Required: true,
-				String: &execution.StringOptionConfig{
-					Default:    " ",
-					TrimSpaces: true,
-				},
-			},
-		},
-		{
-			name: "string option empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeString,
-				Name: "opt",
-			},
-		},
-		{
-			name: "string option with non-string config",
-			option: execution.Option{
-				Type: execution.OptionTypeString,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Format: execution.BoolOptionFormatTrueFalse,
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "select option",
-			option: execution.Option{
-				Type: execution.OptionTypeSelect,
-				Name: "opt",
-				Select: &execution.SelectOptionConfig{
-					Default:     "a",
-					Values:      []string{"a", "b"},
-					AllowCustom: true,
-				},
-			},
-		},
-		{
-			name: "select option empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeSelect,
-				Name: "opt",
-			},
-			wantErr: true,
-		},
-		{
-			name: "select option with non-select config",
-			option: execution.Option{
-				Type: execution.OptionTypeSelect,
-				Name: "opt",
-				Select: &execution.SelectOptionConfig{
-					Default:     "a",
-					Values:      []string{"a", "b"},
-					AllowCustom: true,
-				},
-				String: &execution.StringOptionConfig{},
-			},
-			wantErr: true,
-		},
-		{
-			name: "select option missing values",
-			option: execution.Option{
-				Type:     execution.OptionTypeSelect,
-				Name:     "opt",
-				Required: true,
-				Select:   &execution.SelectOptionConfig{},
-			},
-			wantErr: true,
-		},
-		{
-			name: "select option don't require default value",
-			option: execution.Option{
-				Type:     execution.OptionTypeSelect,
-				Name:     "opt",
-				Required: true,
-				Select: &execution.SelectOptionConfig{
-					Values: []string{"a", "b"},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "select option cannot be empty",
-			option: execution.Option{
-				Type:     execution.OptionTypeSelect,
-				Name:     "opt",
-				Required: true,
-				Select: &execution.SelectOptionConfig{
-					Values:  []string{"a", "b", ""},
-					Default: "a",
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "select option default not in allowed values",
-			option: execution.Option{
-				Type:     execution.OptionTypeSelect,
-				Name:     "opt",
-				Required: true,
-				Select: &execution.SelectOptionConfig{
-					Default:     "c",
-					Values:      []string{"a", "b"},
-					AllowCustom: true,
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "multi option",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-				Multi: &execution.MultiOptionConfig{
-					Delimiter: ",",
-					Values:    []string{"a", "b"},
-					Default:   []string{"a"},
-				},
-			},
-		},
-		{
-			name: "multi option default value not in allowed values",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-				Multi: &execution.MultiOptionConfig{
-					Delimiter:   ",",
-					Values:      []string{"a", "b"},
-					Default:     []string{"b", "c"},
-					AllowCustom: true,
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "multi option empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-			},
-			wantErr: true,
-		},
-		{
-			name: "multi option with non-multi config",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-				Multi: &execution.MultiOptionConfig{
-					Delimiter: ",",
-					Values:    []string{"a", "b"},
-					Default:   []string{"a"},
-				},
-				String: &execution.StringOptionConfig{},
-			},
-			wantErr: true,
-		},
-		{
-			name: "multi option empty delimiter",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-				Multi: &execution.MultiOptionConfig{
-					Values: []string{"a", "b"},
-				},
-			},
-		},
-		{
-			name: "multi option missing values",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-				Multi: &execution.MultiOptionConfig{
-					Delimiter: ",",
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "multi option don't require default",
-			option: execution.Option{
-				Type:     execution.OptionTypeMulti,
-				Name:     "opt",
-				Required: true,
-				Multi: &execution.MultiOptionConfig{
-					Delimiter: ",",
-					Values:    []string{"a", "b"},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "multi option default values not in allowed values",
-			option: execution.Option{
-				Type:     execution.OptionTypeMulti,
-				Name:     "opt",
-				Required: true,
-				Multi: &execution.MultiOptionConfig{
-					Delimiter: ",",
-					Default:   []string{"b", "c"},
-					Values:    []string{"a", "b"},
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "multi option cannot be empty",
-			option: execution.Option{
-				Type:     execution.OptionTypeMulti,
-				Name:     "opt",
-				Required: true,
-				Multi: &execution.MultiOptionConfig{
-					Delimiter: ",",
-					Default:   []string{"b"},
-					Values:    []string{"a", "b", ""},
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "multi default value cannot be empty",
-			option: execution.Option{
-				Type:     execution.OptionTypeMulti,
-				Name:     "opt",
-				Required: true,
-				Multi: &execution.MultiOptionConfig{
-					Delimiter: ",",
-					Default:   []string{""},
-					Values:    []string{"a", "b"},
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "date option",
-			option: execution.Option{
-				Type: execution.OptionTypeDate,
-				Name: "opt",
-				Date: &execution.DateOptionConfig{
-					Format: "YYYY-MM-DD HH:mm:ss",
-				},
-			},
-		},
-		{
-			name: "date option empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeDate,
-				Name: "opt",
-			},
-		},
-		{
-			name: "date option with non-date config",
-			option: execution.Option{
-				Type:   execution.OptionTypeDate,
-				Name:   "opt",
-				String: &execution.StringOptionConfig{},
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			err := options.ValidateJobOption(tt.option, rootPath).ToAggregate()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ValidateJobOption() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
 func TestEvaluateOption(t *testing.T) {
 	tests := []struct {
 		name    string
@@ -1252,8 +795,8 @@ func TestEvaluateOption(t *testing.T) {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			// Validate the option first, as EvaluateOption assumes a valid option.
-			if err := options.ValidateJobOption(tt.option, rootPath).ToAggregate(); err != nil {
-				t.Errorf("ValidateJobOption() got error %v", err)
+			if err := options.ValidateOption(tt.option, rootPath).ToAggregate(); err != nil {
+				t.Errorf("ValidateOption() got error %v", err)
 				return
 			}
 
@@ -1265,274 +808,6 @@ func TestEvaluateOption(t *testing.T) {
 			}
 			if got != tt.want {
 				t.Errorf("EvaluateOption() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestEvaluateOptionDefault(t *testing.T) {
-	tests := []struct {
-		name    string
-		option  execution.Option
-		want    string
-		wantErr bool
-	}{
-		{
-			name: "bool option, missing config",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-			},
-			wantErr: true,
-		},
-		{
-			name: "bool option, default false",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Format: execution.BoolOptionFormatTrueFalse,
-				},
-			},
-			want: "false",
-		},
-		{
-			name: "bool option, default true",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Format:  execution.BoolOptionFormatTrueFalse,
-					Default: true,
-				},
-			},
-			want: "true",
-		},
-		{
-			name: "bool option, default false, custom format",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Format:   execution.BoolOptionFormatCustom,
-					TrueVal:  "--dry-run ",
-					FalseVal: "",
-				},
-			},
-			want: "",
-		},
-		{
-			name: "bool option, default true, custom format",
-			option: execution.Option{
-				Type: execution.OptionTypeBool,
-				Name: "opt",
-				Bool: &execution.BoolOptionConfig{
-					Default:  true,
-					Format:   execution.BoolOptionFormatCustom,
-					TrueVal:  "--dry-run ",
-					FalseVal: "",
-				},
-			},
-			want: "--dry-run ",
-		},
-		{
-			name: "string option, empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeString,
-				Name: "opt",
-			},
-			want: "",
-		},
-		{
-			name: "string option, no default",
-			option: execution.Option{
-				Type:   execution.OptionTypeString,
-				Name:   "opt",
-				String: &execution.StringOptionConfig{},
-			},
-			want: "",
-		},
-		{
-			name: "string option, with default",
-			option: execution.Option{
-				Type: execution.OptionTypeString,
-				Name: "opt",
-				String: &execution.StringOptionConfig{
-					Default: "hello ",
-				},
-			},
-			want: "hello ",
-		},
-		{
-			name: "string option, with default, trim space",
-			option: execution.Option{
-				Type: execution.OptionTypeString,
-				Name: "opt",
-				String: &execution.StringOptionConfig{
-					Default:    "hello ",
-					TrimSpaces: true,
-				},
-			},
-			want: "hello",
-		},
-		{
-			name: "select option, empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeSelect,
-				Name: "opt",
-			},
-			want: "",
-		},
-		{
-			name: "select option, not required, no default",
-			option: execution.Option{
-				Type: execution.OptionTypeSelect,
-				Name: "opt",
-				Select: &execution.SelectOptionConfig{
-					Values: []string{"a", "b"},
-				},
-			},
-			want: "",
-		},
-		{
-			name: "select option, not required, with default",
-			option: execution.Option{
-				Type:     execution.OptionTypeSelect,
-				Name:     "opt",
-				Required: true,
-				Select: &execution.SelectOptionConfig{
-					Default: "a",
-					Values:  []string{"a", "b"},
-				},
-			},
-			want: "a",
-		},
-		{
-			name: "select option, required, no default",
-			option: execution.Option{
-				Type:     execution.OptionTypeSelect,
-				Name:     "opt",
-				Required: true,
-				Select: &execution.SelectOptionConfig{
-					Values: []string{"a", "b"},
-				},
-			},
-			want: "",
-		},
-		{
-			name: "select option, required, with default",
-			option: execution.Option{
-				Type:     execution.OptionTypeSelect,
-				Name:     "opt",
-				Required: true,
-				Select: &execution.SelectOptionConfig{
-					Default: "a",
-					Values:  []string{"a", "b"},
-				},
-			},
-			want: "a",
-		},
-		{
-			name: "multi option, empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-			},
-			want: "",
-		},
-		{
-			name: "multi option, not required, no default",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-				Multi: &execution.MultiOptionConfig{
-					Values:    []string{"a", "b", "c"},
-					Delimiter: ",",
-				},
-			},
-			want: "",
-		},
-		{
-			name: "multi option, not required, with default",
-			option: execution.Option{
-				Type: execution.OptionTypeMulti,
-				Name: "opt",
-				Multi: &execution.MultiOptionConfig{
-					Default:   []string{"c", "b"},
-					Values:    []string{"a", "b", "c"},
-					Delimiter: ",",
-				},
-			},
-			want: "c,b",
-		},
-		{
-			name: "multi option, required, no default",
-			option: execution.Option{
-				Type:     execution.OptionTypeMulti,
-				Name:     "opt",
-				Required: true,
-				Multi: &execution.MultiOptionConfig{
-					Values:    []string{"a", "b", "c"},
-					Delimiter: ",",
-				},
-			},
-			want: "",
-		},
-		{
-			name: "multi option, required, with default",
-			option: execution.Option{
-				Type:     execution.OptionTypeMulti,
-				Name:     "opt",
-				Required: true,
-				Multi: &execution.MultiOptionConfig{
-					Default:   []string{"c", "b"},
-					Values:    []string{"a", "b", "c"},
-					Delimiter: ",",
-				},
-			},
-			want: "c,b",
-		},
-		{
-			name: "date option, empty config",
-			option: execution.Option{
-				Type: execution.OptionTypeDate,
-				Name: "opt",
-			},
-			want: "",
-		},
-		{
-			name: "date option, not required",
-			option: execution.Option{
-				Type: execution.OptionTypeDate,
-				Name: "opt",
-				Date: &execution.DateOptionConfig{
-					Format: "D MMM YYYY",
-				},
-			},
-			want: "",
-		},
-		{
-			name: "date option, required",
-			option: execution.Option{
-				Type: execution.OptionTypeDate,
-				Name: "opt",
-				Date: &execution.DateOptionConfig{
-					Format: "D MMM YYYY",
-				},
-			},
-			want: "",
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := options.EvaluateOptionDefault(tt.option)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("EvaluateOptionDefault() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("EvaluateOptionDefault() got = %v, want %v", got, tt.want)
 			}
 		})
 	}
