+++ pkg/core/options/substitution_test.go
