+++ pkg/core/options/util.go
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2022 The Furiko Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package options
+
+import (
+	"fmt"
+	"reflect"
+
+	"k8s.io/apimachinery/pkg/util/validation/field"
+
+	execution "github.com/furiko-io/furiko/apis/execution/v1alpha1"
+)
+
+// ZeroForNonConfig returns a copy of the Option with all non-config fields
+// reset to the zero value.
+func ZeroForNonConfig(option execution.Option) execution.Option {
+	newOpt := option.DeepCopy()
+	v := reflect.ValueOf(newOpt).Elem()
+	for i := 0; i < v.NumField(); i++ {
+		f := v.Field(i)
+		if f.Kind() != reflect.Ptr {
+			f.Set(reflect.New(f.Type()).Elem())
+		}
+	}
+	return *newOpt
+}
+
+// NewForbiddenOptionConfigError returns a validation error for an OptionType
+// which contains config values other than its own Type.
+func NewForbiddenOptionConfigError(fldPath *field.Path, optionType execution.OptionType) *field.Error {
+	detail := fmt.Sprintf("cannot specify non-%v options for option type %v", optionType, optionType)
+	return field.Forbidden(fldPath, detail)
+}
